Add inter-service communication and service discovery with Netflix Eureka
- Implemented inter-service communication between microservices.
- Integrated Netflix Eureka for service discovery.
- Updated configuration files to include Eureka client settings.
- Added necessary dependencies for Eureka in the `pom.xml`.

+++ order-service/src/main/java/org/projects/orderservice/dto/InstrumentOrderCreateDto.java
@@ -11,7 +11,7 @@ import lombok.Builder;
 public record InstrumentOrderCreateDto(
         @NotNull(message = "Quantity is required")
         @Min(value = 0, message = "Quantity must be greater than 0")
-        Byte quantity,
+        Integer quantity,
 
         @NotNull(message = "Instrument ID is required")
         String instrumentId) {
